Print error message if the adopt api returns one
Sometimes we get an "adoptopenjdk" file back from the adopt api
when we request a binary via curl. This change assumes that's an
error message html file and prints the contents, but only if
we've failed to download after 5 retries.

The reason we retry 5 times is that we already appear to be "trying"
to retry in the event of an error, we just aren't recognising this
error as an error. This is also resolved.

Signed-off-by: Adam Farley <adfarley@redhat.com>